pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "trabfinal_pipeline"
version = "0.1.0"
authors = [
  { name="pedro_cozzati", email="p.cozzati27@gmail.com" },
  { name="yasmin_martins", email="yasmin_martins@gmail.com" },
  { name="marcela_bento", email="marcela_bento@gmail.com" },
  { name="thomas_navajas", email="thomas_navajas@gmail.com" },
]
description = "Um pipeline de dados com PySpark estruturado com boas práticas de engenharia de software."
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyspark==4.0.0",
    "pyyaml==6.0.3"
]

[project.optional-dependencies]
dev = [
    "ruff==0.12.9",
    "black==25.1.0",
    "build==1.3.0"
]

[project.scripts]
run-data-pipeline = "main:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = {find = {where = ["src"]}}
